requirements.txt
django-ninja==1.3.0
django-q2==1.6.2
channels==4.1.0
channels-redis==4.2.0
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
schedule==1.2.2

